+++ components/agenda/agendaDetail/tabs/AgendaAnnouncements.tsx
@@ -1,36 +1,18 @@
-import { useEffect, useState } from 'react';
-import { AnnouncementProps } from 'types/agenda/agendaDetail/announcementTypes';
+import { useRouter } from 'next/router';
 import AnnouncementItem from 'components/agenda/agendaDetail/tabs/AnnouncementItem';
 import { UploadBtn } from 'components/agenda/button/UploadBtn';
+import useAnnouncements from 'hooks/agenda/agendaDetail/useAnnouncements';
 import styles from 'styles/agenda/agendaDetail/tabs/AgendaAnnouncements.module.scss';
-
-const mockData = [
-  {
-    id: 1,
-    title: '첫 번째 공지사항',
-    contents: '공지사항 이렇게 변경합니다 알아서 보세요.',
-    createdAt: new Date(),
-  },
-  {
-    id: 2,
-    title: '두 번째 공지사항',
-    contents: '공지사항 이렇게 변경합니다 알아서 보세요.',
-    createdAt: new Date(),
-  },
-];
 
 const newAnnoucnemet = () => {
   alert('새로운 공지사항을 추가합니다.');
 };
 
 export default function AgendaAnnouncements({ isHost }: { isHost: boolean }) {
-  const [announcementData, setannouncementData] = useState<
-    AnnouncementProps[] | null
-  >(null);
+  const router = useRouter();
+  const { agendaKey } = router.query;
 
-  useEffect(() => {
-    setannouncementData(mockData);
-  }, []);
+  const announcementData = useAnnouncements(agendaKey as string);
 
   if (!announcementData) {
     return <div>Loading...</div>;
